+++ api/src/measurement.rs
@@ -1076,7 +1076,7 @@ mod tests {
                     }, {
                         "atom": "Gram",
                         "prefix": null,
-                        "factor": 1,
+                        "factor": null,
                         "exponent": -4,
                         "annotation": null
                     }]
@@ -1084,7 +1084,7 @@ mod tests {
             }"#.replace("\n", "")
                 .replace(" ", "");
 
-            let term1 = term!(Centi, Meter, factor: 100, exponent: 456, annotation: Some("stuff".to_string()));
+            let term1 = term!(Centi, Meter, factor: 100, exponent: 456, annotation: "stuff".to_string());
             let term2 = term!(Gram, exponent: -4);
 
             let unit = Unit {
@@ -1139,7 +1139,7 @@ mod tests {
 
             let k = serde_json::from_str(json).expect("Couldn't convert JSON String to Unit");
 
-            let term1 = term!(Centi, Meter, factor: 100, exponent: 456, annotation: Some("stuff".to_string()));
+            let term1 = term!(Centi, Meter, factor: 100, exponent: 456, annotation: "stuff".to_string());
             let term2 = term!(Gram, exponent: -4);
 
             let unit = Unit {