+++ api/CHANGELOG.md
@@ -9,12 +9,20 @@
 
 ### Changed
 
+* `Term::factor` and `Term::exponent` are now wrapped in an `Option` (the same
+  goes for their related `parser::terms::mapper` components) to save unnecessary
+  allocations.
 * *Removed* `stdweb` support. This was overkill from the start.
 * Renamed `Decomposable::expression()` to `Decomposable::decompose()` to be more
   conventional.
 * `Decomposable::decompose()` now takes a value to let the caller decide what to
   pass in.
 * `Decomposable` now defines associated types for more `impl` flexibility.
+* Extracted some of `Decomposable::decompose()` to a new method,
+  `Decomposable::format_output()`, which let's consumers customize the formatted
+  output.
+* `Decomposable::numerator()` and `Decomposable::denominator()` now return
+  `Option<String>` instead of `String`.
 
 ## [0.2.0] - 2018-06-26